% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_liquidacao.R
\name{build_liquidacao}
\alias{build_liquidacao}
\title{Cria um data frame com os dados da liquidação.}
\usage{
build_liquidacao(empenho, liquidac)
}
\arguments{
\item{empenho}{Data frame de EMPENHO.TXT obtido por parse_empenho()}

\item{liquidac}{Data frame de LIQUIDAC.TXT obtido por parse_liquidac()}
}
\value{
Um Data Frame com os dados.
}
\description{
Converte os dados gerados para importação pelo SIAPC/PAD do TCE/RS em um Data Frame do R os dados da liquidação.
}
\author{
Everton da Rosa
}
