% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_moviemp.R
\name{build_moviemp}
\alias{build_moviemp}
\title{Cria um data frame com a movimentação dos empenhos.}
\usage{
build_moviemp(empenho, liquidac, pagament)
}
\arguments{
\item{empenho}{Data frame de EMPENHO.TXT obtido por parse_empenho()}

\item{liquidac}{Data frame de LIQUIDAC.TXT obtido por parse_liquidac()}

\item{pagament}{Data frame de PAGAMENT.TXT obtido por parse_pagament()}
}
\value{
Um Data Frame com os dados.
}
\description{
Converte os dados gerados para importação pelo SIAPC/PAD do TCE/RS em um Data Frame do R com a movimentação dos empenhos.
}
\author{
Everton da Rosa
}
