% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_pagamento.R
\name{build_pagamento}
\alias{build_pagamento}
\title{Cria um data frame com os dados dos pagamentos}
\usage{
build_pagamento(empenho, pagament)
}
\arguments{
\item{empenho}{Data frame de EMPENHO.TXT obtido por parse_empenho()}

\item{pagament}{Data frame de PAGAMENT.TXT obtido por parse_pagament()}
}
\value{
Um Data Frame com os dados.
}
\description{
Converte os dados gerados para importação pelo SIAPC/PAD do TCE/RS em um Data Frame do R os dados dos pagamentos.
}
\author{
Everton da Rosa
}
