% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_bal_desp.R, R/parse_empenho.R
\name{parse_bal_desp}
\alias{parse_bal_desp}
\alias{parse_empenho}
\title{Converte BAL_DESP.TXT para um Data Frame.}
\usage{
parse_bal_desp(arquivo_txt)

parse_empenho(arquivo_txt)
}
\arguments{
\item{arquivo_txt}{O caminho para o arquivo EMPENHO.TXT}
}
\value{
Um Data Frame com os dados.

Um Data Frame com os dados.
}
\description{
Converte os dados gerados para importação pelo SIAPC/PAD do TCE/RS em um Data Frame do R.

Converte os dados gerados para importação pelo SIAPC/PAD do TCE/RS em um Data Frame do R.
}
\author{
Everton da Rosa

Everton da Rosa
}
