% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_bal_rec.R
\name{parse_bal_rec}
\alias{parse_bal_rec}
\title{Converte BAL_REC.TXT para um Data Frame.}
\usage{
parse_bal_rec(arquivo_txt)
}
\arguments{
\item{arquivo_txt}{O caminho para o arquivo BAL_REC.TXT}
}
\value{
Um Data Frame com os dados.
}
\description{
Converte os dados gerados para importação pelo SIAPC/PAD do TCE/RS em um Data Frame do R.
}
\author{
Everton da Rosa
}
