% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_bver_enc.R
\name{parse_bver_enc}
\alias{parse_bver_enc}
\title{Converte BVER_ENC.TXT para um Data Frame.}
\usage{
parse_bver_enc(arquivo_txt)
}
\arguments{
\item{arquivo_txt}{O caminho para o arquivo BVER_ENC.TXT}
}
\value{
Um Data Frame com os dados.
}
\description{
Converte os dados gerados para importação pelo SIAPC/PAD do TCE/RS em um Data Frame do R.
}
\author{
Everton da Rosa
}
