% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_bvmovant.R
\name{parse_bvmovant}
\alias{parse_bvmovant}
\title{Converte BVMOVANT.TXT para um Data Frame.}
\usage{
parse_bvmovant(arquivo_txt)
}
\arguments{
\item{arquivo_txt}{O caminho para o arquivo BVMOVANT.TXT}
}
\value{
Um Data Frame com os dados.
}
\description{
Converte os dados gerados para importação pelo SIAPC/PAD do TCE/RS em um Data Frame do R.
}
\author{
Everton da Rosa
}
