% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_credor.R
\name{parse_credor}
\alias{parse_credor}
\title{Converte CREDOR.TXT para um Data Frame.}
\usage{
parse_credor(arquivo_txt)
}
\arguments{
\item{arquivo_txt}{O caminho para o arquivo CREDOR.TXT}
}
\value{
Um Data Frame com os dados.
}
\description{
Converte os dados gerados para importação pelo SIAPC/PAD do TCE/RS em um Data Frame do R.
}
\author{
Everton da Rosa
}
