% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_cta_disp.R
\name{parse_cta_disp}
\alias{parse_cta_disp}
\title{Converte CTA_DISP.TXT para um Data Frame.}
\usage{
parse_cta_disp(arquivo_txt)
}
\arguments{
\item{arquivo_txt}{O caminho para o arquivo CTA_DISP.TXT}
}
\value{
Um Data Frame com os dados.
}
\description{
Converte os dados gerados para importação pelo SIAPC/PAD do TCE/RS em um Data Frame do R.
}
\author{
Everton da Rosa
}
