% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_decreto.R
\name{parse_decreto}
\alias{parse_decreto}
\title{Converte DECRETO.TXT para um Data Frame.}
\usage{
parse_decreto(arquivo_txt)
}
\arguments{
\item{arquivo_txt}{O caminho para o arquivo DECRETO.TXT}
}
\value{
Um Data Frame com os dados.
}
\description{
Converte os dados gerados para importação pelo SIAPC/PAD do TCE/RS em um Data Frame do R.
}
\author{
Everton da Rosa
}
