% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_funcao.R
\name{parse_funcao}
\alias{parse_funcao}
\title{Converte FUNCAO.TXT para um Data Frame.}
\usage{
parse_funcao(arquivo_txt)
}
\arguments{
\item{arquivo_txt}{O caminho para o arquivo FUNCAO.TXT}
}
\value{
Um Data Frame com os dados.
}
\description{
Converte os dados gerados para importação pelo SIAPC/PAD do TCE/RS em um Data Frame do R.
}
\author{
Everton da Rosa
}
