% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_liquidac.R
\name{parse_liquidac}
\alias{parse_liquidac}
\title{Converte LIQUIDAC.TXT para um Data Frame.}
\usage{
parse_liquidac(arquivo_txt)
}
\arguments{
\item{arquivo_txt}{O caminho para o arquivo LIQUIDAC.TXT}
}
\value{
Um Data Frame com os dados.
}
\description{
Converte os dados gerados para importação pelo SIAPC/PAD do TCE/RS em um Data Frame do R.
}
\author{
Everton da Rosa
}
