% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_orgao.R
\name{parse_orgao}
\alias{parse_orgao}
\title{Converte ORGAO.TXT para um Data Frame.}
\usage{
parse_orgao(arquivo_txt)
}
\arguments{
\item{arquivo_txt}{O caminho para o arquivo ORGAO.TXT}
}
\value{
Um Data Frame com os dados.
}
\description{
Converte os dados gerados para importação pelo SIAPC/PAD do TCE/RS em um Data Frame do R.
}
\author{
Everton da Rosa
}
