% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_pagament.R
\name{parse_pagament}
\alias{parse_pagament}
\title{Converte PAGAMENT.TXT para um Data Frame.}
\usage{
parse_pagament(arquivo_txt)
}
\arguments{
\item{arquivo_txt}{O caminho para o arquivo PAGAMENT.TXT}
}
\value{
Um Data Frame com os dados.
}
\description{
Converte os dados gerados para importação pelo SIAPC/PAD do TCE/RS em um Data Frame do R.
}
\author{
Everton da Rosa
}
