% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_programa.R
\name{parse_programa}
\alias{parse_programa}
\title{Converte PROGRAMA.TXT para um Data Frame.}
\usage{
parse_programa(arquivo_txt)
}
\arguments{
\item{arquivo_txt}{O caminho para o arquivo PROGRAMA.TXT}
}
\value{
Um Data Frame com os dados.
}
\description{
Converte os dados gerados para importação pelo SIAPC/PAD do TCE/RS em um Data Frame do R.
}
\author{
Everton da Rosa
}
