% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_projativ.R
\name{parse_projativ}
\alias{parse_projativ}
\title{Converte PROJATIV.TXT para um Data Frame.}
\usage{
parse_projativ(arquivo_txt)
}
\arguments{
\item{arquivo_txt}{O caminho para o arquivo PROJATIV.TXT}
}
\value{
Um Data Frame com os dados.
}
\description{
Converte os dados gerados para importação pelo SIAPC/PAD do TCE/RS em um Data Frame do R.
}
\author{
Everton da Rosa
}
