% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_rd_extra.R
\name{parse_rd_extra}
\alias{parse_rd_extra}
\title{Converte RD_EXTRA.TXT para um Data Frame.}
\usage{
parse_rd_extra(arquivo_txt)
}
\arguments{
\item{arquivo_txt}{O caminho para o arquivo RD_EXTRA.TXT}
}
\value{
Um Data Frame com os dados.
}
\description{
Converte os dados gerados para importação pelo SIAPC/PAD do TCE/RS em um Data Frame do R.
}
\author{
Everton da Rosa
}
