% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_rec_ant.R
\name{parse_rec_ant}
\alias{parse_rec_ant}
\title{Converte REC_ANT.TXT para um Data Frame.}
\usage{
parse_rec_ant(arquivo_txt)
}
\arguments{
\item{arquivo_txt}{O caminho para o arquivo REC_ANT.TXT}
}
\value{
Um Data Frame com os dados.
}
\description{
Converte os dados gerados para importação pelo SIAPC/PAD do TCE/RS em um Data Frame do R.
}
\author{
Everton da Rosa
}
