% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_receita.R
\name{parse_receita}
\alias{parse_receita}
\title{Converte RECEITA.TXT para um Data Frame.}
\usage{
parse_receita(arquivo_txt)
}
\arguments{
\item{arquivo_txt}{O caminho para o arquivo RECEITA.TXT}
}
\value{
Um Data Frame com os dados.
}
\description{
Converte os dados gerados para importação pelo SIAPC/PAD do TCE/RS em um Data Frame do R.
}
\author{
Everton da Rosa
}
