% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_recurso.R
\name{parse_recurso}
\alias{parse_recurso}
\title{Converte RECURSO.TXT para um Data Frame.}
\usage{
parse_recurso(arquivo_txt)
}
\arguments{
\item{arquivo_txt}{O caminho para o arquivo RECURSO.TXT}
}
\value{
Um Data Frame com os dados.
}
\description{
Converte os dados gerados para importação pelo SIAPC/PAD do TCE/RS em um Data Frame do R.
}
\author{
Everton da Rosa
}
