% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_rubrica.R
\name{parse_rubrica}
\alias{parse_rubrica}
\title{Converte RUBRICA.TXT para um Data Frame.}
\usage{
parse_rubrica(arquivo_txt)
}
\arguments{
\item{arquivo_txt}{O caminho para o arquivo RUBRICA.TXT}
}
\value{
Um Data Frame com os dados.
}
\description{
Converte os dados gerados para importação pelo SIAPC/PAD do TCE/RS em um Data Frame do R.
}
\author{
Everton da Rosa
}
