% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_tce_4111.R
\name{parse_tce_4111}
\alias{parse_tce_4111}
\title{Converte tce_4111.TXT para um Data Frame.}
\usage{
parse_tce_4111(arquivo_txt)
}
\arguments{
\item{arquivo_txt}{O caminho para o arquivo tce_4111.TXT}
}
\value{
Um Data Frame com os dados.
}
\description{
Converte os dados gerados para importação pelo SIAPC/PAD do TCE/RS em um Data Frame do R.
}
\author{
Everton da Rosa
}
