% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_uniorcam.R
\name{parse_uniorcam}
\alias{parse_uniorcam}
\title{Converte UNIORCAM.TXT para um Data Frame.}
\usage{
parse_uniorcam(arquivo_txt)
}
\arguments{
\item{arquivo_txt}{O caminho para o arquivo UNIORCAM.TXT}
}
\value{
Um Data Frame com os dados.
}
\description{
Converte os dados gerados para importação pelo SIAPC/PAD do TCE/RS em um Data Frame do R.
}
\author{
Everton da Rosa
}
